bot/pyproject.toml
[project]
name = "lms-bot"
version = "0.1.0"
description = "Telegram bot for the LMS"
requires-python = ">=3.11"
dependencies = [
  "python-telegram-bot>=21.0",
  "httpx>=0.27.0",
  "python-dotenv>=1.0.0",
]
